Use fingerprint to detect changes in methods that are intrinsified. Make sure that the fingerprints are portable across 32 and 64 bit platforms. Review URL: https://codereview.chromium.org//11421024
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@15217 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/object.h
@@ -1438,7 +1438,7 @@ class Function : public Object {
 
   // Slow function, use in asserts to track changes in important library
   // functions.
-  intptr_t SourceFingerprint() const;
+  int32_t SourceFingerprint() const;
 
   static const int kCtorPhaseInit = 1 << 0;
   static const int kCtorPhaseBody = 1 << 1;
